fix: duration is double

+++ src/main/kotlin/com/github/hank9999/kook/types/Attachments.kt
@@ -11,7 +11,7 @@ data class Attachments(
     @Required val name: String = "",
     @SerialName("file_type") val fileType: String = "",
     val size: Int = 0,
-    val duration: Int = 0,
+    val duration: Double = 0.0,
     val width: Int = 0,
     val height: Int = 0
 )
